+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Upcoming
+## Added
+- Support for request attributes from katana 1.1.12
+- Accessors for request id and timestamp
+
 ## [1.1.8] - 2017-07-13
 ## Added
 - New header methods